#139 fix scrolling issue
