+++ packages/web/package.json
@@ -10,6 +10,7 @@
         "bootstrap": "^5.1.3",
         "eslint-config-react-app": "^7.0.1",
         "framer-motion": "^6.3.11",
+        "jquery": "^3.6.3",
         "react": "^18.2.0",
         "react-bootstrap": "^2.4.0",
         "react-dom": "^18.2.0",
