+++ packages/web/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import LandingHome from './components/spa/MainPage';
-import PageNotFound from './components/other/pageNotFound';
+import NotFound from './components/other/NotFound';
 import LandingAdminPage from './components/admin_page/landing_admin_page';
 import Dashboard from './components/admin_page/admin_dashboard';
 import MemberActions from './components/admin_page/members_page/single_member';
@@ -85,14 +85,8 @@ function App() {
                 path="/admin/dashboard/mentors/actions"
                 element={<MentorsActions />}
             />
-            <Route 
-                path="/admin/dashboard/jury"
-                element={<RenderJury />} 
-            />
-            <Route 
-                path="/admin/dashboard/jury/add"
-                element={<AddJury />} 
-            />
+            <Route path="/admin/dashboard/jury" element={<RenderJury />} />
+            <Route path="/admin/dashboard/jury/add" element={<AddJury />} />
             <Route
                 path="/admin/dashboard/jury/actions"
                 element={<JuryActions />}
@@ -121,14 +115,8 @@ function App() {
                 path="/admin/dashboard/partners/actions"
                 element={<PartnersActions />}
             />
-            <Route 
-                path="/jobs"
-                element={<JobsSection />}
-            />
-            <Route
-                path="/*"
-                element={<PageNotFound/>}
-            />
+            <Route path="/jobs" element={<JobsSection />} />
+            <Route path="/*" element={<NotFound />} />
         </Routes>
     );
 }
