--- packages/web/src/components/other/pageNotFound.css
@@ -1,14 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400;500;600;700&display=swap');
-
-.background {
-    background: linear-gradient(rgb(21, 76, 121), rgb(118, 181, 197));
-    height: 100vh;
-    width: 100vw;
-}
-
-.hubzie-404-image {
-    max-width: 100%;
-    height: auto;
-    display: table;
-    margin: 0 auto;
-}
