+++ packages/web/src/components/spa/MainPage.jsx
@@ -5,7 +5,8 @@ import { NavBar } from './Navigation/NavBar';
 import { LandingSection } from './LandingSection/LandingSection';
 import { Anchor, Props } from './Navigation/NavFactory.js';
 import { ArticlesSection } from './ArticlesSection/ArticlesSection';
-//import { useEffect } from 'react';
+import { useEffect } from 'react';
+// import $ from 'jquery';
 
 const LandingHome = () => {
     const anchorList = [
@@ -16,13 +17,26 @@ const LandingHome = () => {
         new Anchor('Jobs', 'jobs')
     ];
 
-    /*eslint-disable no-self-assign*/
-    /*
-        useEffect(()=> {
-        console.log(!!location.hash)
-        if(location.hash) 
-        setTimeout(()=> location.hash = location.hash,1000)
-    },[])*/
+    useEffect(() => {
+        let hasHash = !!location.hash;
+        if (hasHash) {
+            /*
+                check if document.getElementById is null
+                -> if null -> skip
+                -> if not null -> trigger .scrollIntoView()
+
+                ---
+
+                try to fix not scrolling into full view
+            */
+
+            setTimeout(() => {
+                document
+                    .getElementById(location.hash.replace('#', ''))
+                    .scrollIntoView();
+            }, 600);
+        }
+    }, []);
 
     return (
         <div className="main">
